+++ chaoslib/types.py
@@ -5,7 +5,7 @@ __all__ = ["MicroservicesStatus", "Probe", "Action", "Experiment", "Layer",
            "TargetLayers", "Activity", "Journal", "Run", "Secrets", "Step",
            "Configuration", "Discovery", "DiscoveredActivities", "Extension",
            "DiscoveredSystemInfo", "Settings", "EventPayload", "Tolerance",
-           "Hypothesis", "Control"]
+           "Hypothesis", "Control", "ConfigVars", "SecretVars"]
 
 
 Action = Dict[str, Any]
@@ -37,3 +37,6 @@ Tolerance = Union[int, str, bool, list, Dict[str, Any]]
 Extension = Dict[str, Any]
 Hypothesis = Dict[str, Any]
 Control = Dict[str, Any]
+
+ConfigVars = Dict[str, Any]
+SecretVars = Secrets